+++ app/src/main/java/site/anzz/childkiosk/util/AmapLocationProvider.kt
@@ -37,3 +37,7 @@ interface AmapLocationProvider {
     fun stopAssistantLocation(webView: WebView)
     fun stopAllAssistantLocations()
 }
+
+interface AmapLocationApiKeyConfigurator {
+    fun configureApiKey(apiKey: String)
+}